backend/requirements/base-requirements.txt
# Base Requirements for all Enviornments

# Django related dependancies
django==5.2.7
django-environ==0.12.0
psycopg2-binary==2.9.10

# REST API framework
djangorestframework==3.15.2
